Adds valid decode function! \o/
This one was particularly hard and took me a long time before I internalized
that filter would be my answer to parse the substitution chart while inside a
loop. I tried many loops inside loops and fors and ifs... things got really
messy.

At this point in my studies I still work with the Cheat Sheet in full-screen
behind my Emacs window, so I struggled to navigate the map and get stuff out of
it because it isn't clear at first that maps are "sequable" and I kept looking
at the functions for maps. When I realized that, I looked into the functions
provided to examine sequences and voila, filter is front and center.

After that, I understood that I already had the correct model for the decode
function in the encode function! :) So I already had the right loop (through the
letters of the strings) and realized that the exterior function to get the
decoded character would be the right place to use filter to parse the chart at
every combination. Here I already understood that I had a mental model that
would work and with a bit of time at the repl it all came together.

I still feel that the code isn't beautiful, that I don't like those if functions
there and that I really need to read more code to understand how to name my
variables and functions in a sane way. But it works!